% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arPLS.R
\docType{data}
\name{Abelsonite}
\alias{Abelsonite}
\title{Raw Raman spectrum for Abelsonite}
\format{
An object of class \code{data.frame} with 3315 rows and 2 columns.
}
\source{
\url{https://rruff.info/repository/sample_child_record_raman_full/by_minerals/Abelsonite__R070007__Broad_Scan__532__0__unoriented__Raman_Data_RAW__13756.txt}
}
\usage{
Abelsonite
}
\description{
A data frame containing 3315 rows and 2 variables (wavenumber and measurement)
}
\examples{
data("Abelsonite")
}
\references{
Lafuente B, Downs R T, Yang H, Stone N (2015) The power of databases: the RRUFF project.
In: Highlights in Mineralogical Crystallography,
T Armbruster and R M Danisi, eds. Berlin, Germany, W. De Gruyter, pp 1-30
}
\author{
Bob Downs \email{rdowns@u.arizona.edu}
}
\keyword{datasets}
