% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arPLS.R
\name{baseline_correction}
\alias{baseline_correction}
\title{asymmetrically reweighted penalized least squares}
\usage{
baseline_correction(
  y,
  lambda = 1e+06,
  ratio = 1e-06,
  max_iter = 50,
  verbose = FALSE,
  cpp = TRUE
)
}
\arguments{
\item{y}{Numeric vector representing the spectrum.}

\item{lambda}{Smoothing parameter. The smaller the more curvature (wiggliness). (default: 1e6).}

\item{ratio}{Stopping criterion based on changes in weight vector per iteration (default: 1e-6).}

\item{max_iter}{Maximum number of iterations as fall back criterion if no conversion happens (default: 50).}

\item{verbose}{Boolean to print intermediary outputs (default: FALSE).}
}
\value{
object of class arPLSresult:
\itemize{
  \item \code{rawinput}: The original spectrum fed into the algorithm.
  \item \code{lambda}: The lambda parameter fed into the algorithm.
  \item \code{ratio}: The ratio stopping parameter fed into the algorithm.
  \item \code{max_iter}: The maximum iteration stopping parameter fed into the algorithm.
  \item \code{baseline}: The fitted spectral baseline.
  \item \code{last_iter}: The number of iterations the algorithm did before stopping.
  \item \code{last_ratio}: The last value of the ratio stopping criterium before stopping.
}
}
\description{
Baseline correction using asymmetrically reweighted penalized least
squares smoothing (Baek et al. 2015).
}
\details{
The algorithm iteratively estimates a spectral baseline curve by updating a
weight vector by means of a generalized logistic function that focuses the
estimation efforts on regions where the baseline and the signal are close to
each other
}
\examples{
{
y <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
baseline <- baseline_correction(y)
}
}
\references{
Baek, S.-J., Park, A., Ahn, Y.-J., and Choo, J. (2015). Baseline correction
using asymmetrically reweighted penalized least squares smoothing.
Analyst, 140:250–257.
}
\author{
Corvin Idler
}
