% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_armaInv}
\alias{rcpparma_armaInv}
\title{Title Armadillo package matrix inversion function
Description Takes a matrix and inverts it.}
\usage{
rcpparma_armaInv(x)
}
\arguments{
\item{x}{matrix to be inverted}
}
\value{
Inverted matrix
}
\description{
Title Armadillo package matrix inversion function
Description Takes a matrix and inverts it.
}
