% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arPLS.R
\name{summary.arPLSresult}
\alias{summary.arPLSresult}
\title{Take an object of class arPLSresult and summarize (print) some facts about it}
\usage{
\method{summary}{arPLSresult}(x)
}
\arguments{
\item{x}{A result object of class arPLSresult (mainly a list).}
}
\description{
This is an S3 generic. To summarize (print) some facts about
the arPLS baseline estimation that led to it
}
