% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arPLS.R
\name{summary.arPLSresult}
\alias{summary.arPLSresult}
\title{Take an object of class arPLSresult and summarize (print) some facts about it}
\usage{
\method{summary}{arPLSresult}(object, ...)
}
\arguments{
\item{object}{A result object of class arPLSresult (mainly a list).}

\item{...}{placeholder for arbitrary additional parameters (to stay in line with other generic summary functions)}
}
\description{
This is an S3 generic.
To summarize (print) some facts about the arPLS baseline estimation that led to it.
}
\examples{
{
data("Abelsonite")
baseline <- baseline_estimation(Abelsonite$measurement, max_iter=10, verbose=T)
summary(baseline)
}
}
